EDIT: Major changes Added nested routing for all the necesarry pages Please pay attention before making a PR
