+++ src/Pages/AllVacancies/AllVacancies.scss
