+++ src/Pages/EmployerPage/CompanyMessages/CompanyMessage.scss
@@ -1,4 +1,4 @@
-@use "../../App.scss" as styles;
+@use '../../../App.scss' as styles;
 
 .cabinet-message-wrapper {
   width: 100%;
