+++ src/Pages/EmployerPage/CompanyVacancies/ChooseMethodToCreate.jsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { useState } from 'react';
 import { Link } from 'react-router-dom';
 
-import Header from '../../Components/Header/Header';
-import Nav from '../../Components/Nav/Nav';
-import SolidButton from '../../Components/Buttons/SolidButton/SolidButton';
-import InputField from '../../Components/Input/InputField/InputField';
-import SelectField from '../../Components/Input/InputSelect/InputSelect';
+import Header from '../../../Components/Header/Header';
+import Nav from '../../../Components/Nav/Nav';
+import SolidButton from '../../../Components/Buttons/SolidButton/SolidButton';
+import InputField from '../../../Components/Input/InputField/InputField';
+import SelectField from '../../../Components/Input/InputSelect/InputSelect';
 
 import './ChooseMethodToCreate.scss';
 /**
