+++ src/Pages/EmployerPage/CompanyVacancies/ChooseMethodToCreate.scss
@@ -1,4 +1,4 @@
-@use '../../App.scss' as styles;
+@use '../../../App.scss' as styles;
 
 .choose-vacancy-wrapper {
     height: 100%;
