+++ src/Pages/EmployerPage/CompanyVacancies/CompanyVacancies.scss
@@ -1,4 +1,4 @@
-@use '../../App.scss' as styles;
+@use '../../../App.scss' as styles;
 .vacancies-wrapper{
     width: 100%;
     height: 100%;
