+++ src/Pages/EmployerPage/CompanyVacancies/ConfirmCreatedVacancy.scss
@@ -1,4 +1,4 @@
-@use '../../App.scss' as styles;
+@use '../../../App.scss' as styles;
 .confirm-wrapper{
     width: 100vw;
     height: max-content;
