+++ src/Pages/EmployerPage/CompanyVacancies/CreateVacancy.scss
@@ -1,4 +1,4 @@
-@use '../../App.scss' as styles;
+@use '../../../App.scss' as styles;
 
 .create-vacancy-wrapper {
     width: 100%;
