+++ src/Pages/EmployerPage/CompanyVacancies/VacancyCreated.scss
@@ -1,4 +1,4 @@
-@use '../../App.scss' as styles;
+@use '../../../App.scss' as styles;
 .vacancy-created-wrapper{
     width: 100%;
     height: max-content;
