+++ src/Pages/SeekerPage/MessagesAndEvents/SeekerResumePage.scss
@@ -1,4 +1,4 @@
-@use "../../App.scss" as styles;
+@use "../../../App.scss" as styles;
 
 .resume-page-wrapper {
   height: 100%;
